Move cpu only stats stuff to xsf/cpu/stats.h
